Extract readiness behaviour.
